Firewall
In the following code we have a firewall that will filter any AUTH
command sent to the vulnerable service but allow all others through.
However firewall uses a packet filtering instead of stateful stream
reassembly , that   is the firewall will match its filtering predicates
against payloads but not against the entire contents of the network
stream. So  this code provides a way to bypass the firewall to pass the
AUTH  command.