Hash-Chains
Here we have been given a situation where the insider has leaked three
password hashes for system account on our given server. We need to
recover the corresponding password using the precomputed table of hash
chains and using  that password we need to login into system and recover
a secret key present in the server.
We have been given the salt for each three account and length of each
chains. The hash function is MD5 and reduction R for a hash value of
length is given by a mathematical model.